# apps/api/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/auth/login": {
      "post": {
        "operationId": "login",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Signed in",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LoginResponseEntity"
                }
              }
            }
          },
          "400": {
            "description": "Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "summary": "Exchange a verified Azure AD id_token for the API's own access/refresh tokens",
        "tags": [
          "Auth"
        ]
      }
    },
    "/auth/register": {
      "post": {
        "operationId": "register",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Registered and signed in",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LoginResponseEntity"
                }
              }
            }
          },
          "400": {
            "description": "Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "summary": "Create (or link) an email+password account",
        "tags": [
          "Auth"
        ]
      }
    },
    "/auth/login-password": {
      "post": {
        "operationId": "loginPassword",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PasswordLoginDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Signed in",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LoginResponseEntity"
                }
              }
            }
          },
          "400": {
            "description": "Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "summary": "Sign in with email+password",
        "tags": [
          "Auth"
        ]
      }
    },
    "/auth/refresh": {
      "post": {
        "operationId": "refresh",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RefreshDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Refreshed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RefreshResponseEntity"
                }
              }
            }
          },
          "400": {
            "description": "Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "summary": "Mint a new access token from a refresh token",
        "tags": [
          "Auth"
        ]
      }
    },
...
  },
  "info": {
    "title": "Linktal Recruitment API",
    "description": "API for Linktal Recruitment OS",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
      "LoginDto": {
        "type": "object",
        "properties": {
          "idToken": {
            "type": "string",
            "description": "Azure AD id_token from the web app's OAuth sign-in"
          }
        },
        "required": [
          "idToken"
        ]
      },
      "SessionUserEntity": {
        "type": "object",
        "properties": {
          "consultantId": {
            "type": "string"
          },
          "email": {
            "type": "object"
          },
          "fullName": {
            "type": "string"
          },
          "roleName": {
            "type": "string",
            "nullable": true
          },
          "permissions": {
            "description": "Flat 'resource:action' strings",
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "industryIds": {
            "description": "Industry ids this consultant is scoped to (only enforced for roleName === \"consultant\")",
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        },
        "required": [
          "consultantId",
          "email",
          "fullName",
          "roleName",
          "permissions",
          "industryIds"
        ]
      },
      "LoginResponseEntity": {
        "type": "object",
        "properties": {
          "accessToken": {
            "type": "string"
          },
          "accessTokenExpiresAt": {
            "type": "number",
            "description": "Epoch ms"
          },
          "refreshToken": {
            "type": "string"
          },
          "user": {
            "$ref": "#/components/schemas/SessionUserEntity"
          }
        },
        "required": [
          "accessToken",
          "accessTokenExpiresAt",
          "refreshToken",
          "user"
        ]
      },
      "RegisterDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string"
          },
          "password": {
            "type": "string",
            "minLength": 8
          },
          "fullName": {
            "type": "string"
          }
        },
        "required": [
          "email",
          "password",
          "fullName"
        ]
      },
      "PasswordLoginDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string"
          },
          "password": {
            "type": "string"
          }
        },
        "required": [
          "email",
          "password"
        ]
      },
      "RefreshDto": {
        "type": "object",
        "properties": {
          "refreshToken": {
            "type": "string",
            "description": "Refresh token issued by POST /auth/login"
          }
        },
        "required": [
          "refreshToken"
        ]
      },
      "RefreshResponseEntity": {
        "type": "object",
        "properties": {
          "accessToken": {
            "type": "string"
          },
          "accessTokenExpiresAt": {
            "type": "number",
            "description": "Epoch ms"
          },
          "user": {
            "$ref": "#/components/schemas/SessionUserEntity"
          }
        },
        "required": [
          "accessToken",
          "accessTokenExpiresAt",
          "user"
        ]
      },
...
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "statusCode": {
            "type": "number",
            "example": 404
          },
          "code": {
            "type": "string",
            "example": "NOT_FOUND",
            "description": "Stable machine-readable code"
          },
          "message": {
            "type": "string",
            "example": "Candidate abc not found"
          },
          "details": {
            "nullable": true,
            "description": "Field-level messages (e.g. validation errors)",
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "path": {
            "type": "string",
            "example": "/api/candidates/abc"
          },
          "timestamp": {
            "type": "string",
            "example": "2026-07-09T10:20:30.000Z"
          }
        },
        "required": [
          "statusCode",
          "code",
          "message",
          "details",
          "path",
          "timestamp"
        ]
      }
    }
  }
}